requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos[dbt-postgres]
apache-airflow-providers-postgres
pandas
SQLAlchemy
psycopg2-binary